Add composite build for Plugin module
The Gradle settings were updated to include a composite build definition.
This allows the Plugin module, located in a separate directory, to be
included as a dependency in the TestAndroidJetpack project.

The `gradle.xml` file now specifies the path to the Plugin module and
configures the IDE to recognize it as part of the project structure.